feat(worker): add cityGmlAttributes to Feature attributes in citygml.rs and citygml.rs (#329)
The code changes add the "cityGmlAttributes" attribute to the Feature
struct in the citygml.rs and citygml.rs files. This attribute is
populated with the attributes of the root entity in the CityGML file.

Based on the recent user commits and repository commits, the established
convention for commit messages in this repository is to use a prefix
indicating the type of change (e.g., feat for a new feature, fix for a
bug fix, chore for maintenance tasks). The commit messages are concise
and provide a clear description of the code changes.

Please note that the commit message provided above is a suggestion based
on the code changes and established conventions. Feel free to modify it
as needed.